git-svn-id: http://xe-core.googlecode.com/svn/sandbox@6640 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ modules/wiki/wiki.model.php
@@ -26,8 +26,7 @@
             header("Pragma: no-cache");
 
             $cache_file = sprintf('%sfiles/cache/wiki/%d.xml', _XE_PATH_,$this->module_srl);
-            //if(!file_exists($cache_file)) FileHandler::writeFile($cache_file, $this->loadWikiTreeList($this->module_srl));
-            FileHandler::writeFile($cache_file, $this->loadWikiTreeList($this->module_srl));
+            if(!file_exists($cache_file)) FileHandler::writeFile($cache_file, $this->loadWikiTreeList($this->module_srl));
 
             print FileHandler::readFile($cache_file);
             Context::close();
